+++ src/lsp/cobol_parser/grammar_utils.ml
@@ -24,13 +24,13 @@ module Overlay_manager =
     let name = __MODULE__
   end) ()
 
-let relation_condition ~neg (binrel: binary_relation) = function
-  | None ->
-      Cobol_ptree.Terms_helpers.neg_condition ~neg @@ Relation binrel
-  | Some (LOr, flatop) ->
-      Abbrev (neg, binrel, LOr, flatop)
-  | Some (LAnd, flatop) ->
-      Abbrev (neg, binrel, LAnd, flatop)
+let relation_condition (abbrev_rel: abbrev_combined_relation) =
+  match abbrev_rel with
+  | neg, subject, AbbrevRelOp (op, AbbrevObject (obj_neg, obj)) ->
+      let neg = if neg then not obj_neg else obj_neg in
+      Cobol_ptree.Terms_helpers.neg_condition ~neg @@ Relation (subject, op, obj)
+  | _ ->
+      Abbrev abbrev_rel
 
 
 type data_division_sentence =
@@ -94,3 +94,24 @@ let build_data_division = function
               { d with screen_sections = ss :: d.screen_sections }
       in
       Some (rebuild_sections list empty_data_division &@ loc)
+
+let build_simple_program opts_par env_div datat_div proc_div ~pos =
+  Program {
+    program_name =
+      (Cobol_ptree.Dummies.dummy_name_or_literal ~pos) &@ Cobol_common.Srcloc.dummy;
+      (* TODO: Use filename *)
+    program_as = None;
+    program_level = ProgramDefinition {
+      mode = None;
+      has_identification_division_header = false;
+      preliminary_informational_paragraphs = [];
+      supplementary_informational_paragraphs = [];
+      nested_programs = []; };
+    program_options = opts_par;
+    program_env = env_div;
+    program_data = (match datat_div with
+      | Some dd -> build_data_division dd
+      | None -> None);
+    program_proc = Some proc_div;
+    program_end_name = None;
+  }