+++ src/lsp/cobol_preproc/preproc_directives.ml
@@ -15,7 +15,7 @@ open Cobol_common.Srcloc.TYPES
 open Text.TYPES
 
 type compiler_directive =
-  | CDir_source of Src_format.any with_loc
+  | CDir_source_format of Src_format.any with_loc
   | CDir_control_section                       (* SECTION in CONTROL DIVISION *)
   | CDir_preproc of Compdir_tree.preproc_directive