+++ src/vscode/superbol-vscode-lib/superbol_vscode_lib.mli
@@ -29,3 +29,4 @@ val deactivate
 module Printer = Superbol_printer
 module Instance = Superbol_instance
 module Workspace = Superbol_workspace
+module Extension = Superbol_extension